core/models/tag.py
from django.db import models
from django.utils.translation import ugettext_lazy as _
from model_utils.models import TimeStampedModel


class Tag(TimeStampedModel):
    title = models.CharField(max_length=255, verbose_name='Название')

    class Meta:
        verbose_name = _('Тег')
        verbose_name_plural = _('Теги')

    def __str__(self):
        return self.title
