requirements.txt
Django==3.0.3
django-choices==1.7.1
django-extensions==2.2.8
django-model-utils==4.0.0
factory-boy==2.12.0
psycopg2==2.8.4
pytest==5.3.5
pytest-cov==2.8.1
pytest-django==3.8.0
pytest-mock==2.0.0
python-magic==0.4.15
